FIX: don't attempt to save child object before parent is saved. (#22)
Previously, when we used the `find_or_create_by` command, it tried to save the child object even before the parent object was saved. It returned the error "You cannot call create unless the parent is saved". To prevent it, we're going to use the `find_or_initialize_by` method instead.