PERF: fetch fields and field options on-demand while creating an issue.
Previously, we stored all the combinations of issue types and fields in the DB. Since most of the fields and options will be used in multiple projects it created unnecessary duplicate data. Also, it reduced the performance of the DB.

Now we're only storing the Jira projects and issue types in database. We will load the other data on-demand while user creating the issue from front-end.