SECURITY: use `FinalDestination` to prevent SSRF attacks on API calls. (#50)
Also, it will check the issue key value on the "attach" action and won't allow the `/` character in it.

Previously, an administrator could point the Jira server to an arbitrary location. After entering an issue key or an arbitrary path in the “Attach Issue” feature, the request failed, however, the output of the response was reflected in the error logs. This vulnerability could be used to enumerate ports on the remote host locally, as well as retrieve responses from non-public-facing services, such as the Prometheus Ruby Exporter. While the path to the issue lookup was automatically appended by the application, it was possible to use the “…/” notation to manipulate the path from the interface. Therefore, any user with the ability to add an existing issue to a POST could perform arbitrary GET requests against the Jira API, however, only the administrator could see the output in the error logs.

Co-authored-by: David Taylor <david@taylorhq.com>