FIX: Enqueue sidekiq job to sync jira projects on jira server instance (#116)
Jira Server doesn't support `issueTypes` in the `expand` query param
when fetching projects. As a result, we were working around the problem
by making one additional API call to fetch a project's `issueTypes`.

This will however only work when there is no rate limiting in place.
When the jira server is configured with a rate limit and there are alot
of projects to sync, the sync operation will end up hitting the rate
limit and also cause other API operations to hit the rate limit as well.

This commit fixes the problem by enqueuing a sidekiq job, with some
jitter, to sync projects. By spreading out the sync over a longer period
of time, it will decrease the chances of running into rate limit
problems.

Related links:

https://community.atlassian.com/forums/Jira-questions/Does-JIRA-On-Premise-support-expanding-issueTypes-in-project-api/qaq-p/774736